load_test/requirements.txt
locust==2.17.0
pillow==10.0.1
requests==2.31.0